fix(static/frontend): update search aria label
Currently the search input aria-label is static and
when focused by a screenreader, the screenreader does
not read the search tip found below the input. This
change checks the .TipIndex value and if it matches
adds the id=SearchTipContent which the input's
aria-describedby is attached to, thus allowing the
visible tip/description to be read after the input's
existing aria-label.

Fixes b/#283297190